Make printing method able to print Admins and Clients
