+++ src/main/java/Model/UserAbstract.java
@@ -55,4 +55,17 @@ public abstract class UserAbstract {
         this.password = password;
     }
 
+    @Override
+    public String toString() {
+        String returnValue;
+        returnValue = String.format("| %s | %s | %s | %s | %s |",
+                this.getID(),
+                this.getName(),
+                this.getSurname(),
+                this.getLogin(),
+                this.getPassword());
+
+        return returnValue;
+    }
+
 }
