+++ src/main/java/View/AbstractView.java
@@ -1,11 +1,12 @@
 package View;
 
 import Model.Admin;
+import Model.UserAbstract;
 
 import java.util.List;
 
 public abstract class AbstractView {
     public abstract void  print(String message);
-    public abstract void print(List<Admin> AdminsList);
+    public abstract void print(List<UserAbstract> PersonsList);
 
 }
