+++ src/main/java/View/TerminalView.java
@@ -1,8 +1,10 @@
 package View;
 
 import Model.Admin;
+import Model.UserAbstract;
 import View.AbstractView;
 
+import java.sql.SQLOutput;
 import java.util.List;
 
 public class TerminalView extends AbstractView {
@@ -13,8 +15,9 @@ public class TerminalView extends AbstractView {
     }
 
     @Override
-    public void print(List<Admin> AdminsList) {
-        for (Admin person: AdminsList
+    public void print(List<UserAbstract> PersonsList) {
+        System.out.println("| ID | First Name | Last Name | Login | Password |\n");
+        for (UserAbstract person: PersonsList
         ) {
             System.out.println(person.toString());
         }
